@@ -2,7 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//+build windows plan9
+//go:build windows || plan9
+// +build windows plan9
 
 package logging
 
