@@ -154,17 +154,18 @@ type stringFormatter struct {
 // The verbs:
 //
 // General:
-//     %{id}        Sequence number for log message (uint64).
-//     %{pid}       Process id (int)
-//     %{time}      Time when log occurred (time.Time)
-//     %{level}     Log level (Level)
-//     %{module}    Module (string)
-//     %{program}   Basename of os.Args[0] (string)
-//     %{message}   Message (string)
-//     %{longfile}  Full file name and line number: /a/b/c/d.go:23
-//     %{shortfile} Final file name element and line number: d.go:23
-//     %{callpath}  Callpath like main.a.b.c...c  "..." meaning recursive call ~. meaning truncated path
-//     %{color}     ANSI color based on log level
+//
+//	%{id}        Sequence number for log message (uint64).
+//	%{pid}       Process id (int)
+//	%{time}      Time when log occurred (time.Time)
+//	%{level}     Log level (Level)
+//	%{module}    Module (string)
+//	%{program}   Basename of os.Args[0] (string)
+//	%{message}   Message (string)
+//	%{longfile}  Full file name and line number: /a/b/c/d.go:23
+//	%{shortfile} Final file name element and line number: d.go:23
+//	%{callpath}  Callpath like main.a.b.c...c  "..." meaning recursive call ~. meaning truncated path
+//	%{color}     ANSI color based on log level
 //
 // For normal types, the output can be customized by using the 'verbs' defined
 // in the fmt package, eg. '%{id:04d}' to make the id output be '%04d' as the
@@ -191,11 +192,12 @@ type stringFormatter struct {
 // future.
 //
 // Experimental:
-//     %{longpkg}   Full package path, eg. github.com/go-logging
-//     %{shortpkg}  Base package path, eg. go-logging
-//     %{longfunc}  Full function name, eg. littleEndian.PutUint32
-//     %{shortfunc} Base function name, eg. PutUint32
-//     %{callpath}  Call function path, eg. main.a.b.c
+//
+//	%{longpkg}   Full package path, eg. github.com/go-logging
+//	%{shortpkg}  Base package path, eg. go-logging
+//	%{longfunc}  Full function name, eg. littleEndian.PutUint32
+//	%{shortfunc} Base function name, eg. PutUint32
+//	%{callpath}  Call function path, eg. main.a.b.c
 func NewStringFormatter(format string) (Formatter, error) {
 	var fmter = &stringFormatter{}
 
